实现旅游社后端 entity+mapper+service+controller 编码

+++ tourism-manage-system/src/main/java/com/swu/tourismmanagesystem/entity/agency/VehicleRealLocation.java
@@ -1,12 +1,13 @@
 package com.swu.tourismmanagesystem.entity.agency;
 import lombok.Data;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 @Data
 public class VehicleRealLocation {
     private Long id;
     private Long vehicleId;
-    private Double longitude;  // 经度
-    private Double latitude;   // 纬度
+    private BigDecimal longitude;  // 经度
+    private BigDecimal latitude;   // 纬度
     private Integer speed;
     private String vehicleStatus;
     private LocalDateTime updateTime;
